backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlmodel==0.0.14
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
# Optional: psycopg2-binary==2.9.9  # Only needed for PostgreSQL

